♻️  [Refactor]: Add routeHeaderInfo to constant
기존의 routePaths.ts에 route path만 저장했지만, route에 따른 Header의 정보를 추가로 작성.
이에 따라 routePaths.ts -> route.ts로 파일명 변경
