+++ src/components/organisms/HouseCard.tsx
@@ -1,4 +1,4 @@
-import routePaths from '@/constants/routePaths';
+import { routePaths } from '@/constants/route';
 import Link from '@/components/atoms/Link';
 import Container from '@/components/atoms/Container';
 import Img from '@/components/atoms/Img';
