+++ src/components/pages/About.tsx
@@ -1,6 +1,6 @@
 import { useRecoilValue } from 'recoil';
 import { useNavigate } from 'react-router-dom';
-import routePaths from '@/constants/routePaths';
+import { routePaths } from '@/constants/route';
 import { supabase } from '@/libs/supabaseClient';
 import Container from '@/components/atoms/Container';
 import Button from '@/components/atoms/Button';
