+++ src/components/templates/MyActivity.template.tsx
@@ -22,7 +22,7 @@ import BadgeIcon from '@/components/molecules/BadgeIcon';
 import Button from '@/components/atoms/Button';
 import useModal from '@/hooks/useModal';
 import { ProfileModifyModalState } from '@/types/modal.type';
-import routePaths from '@/constants/routePaths';
+import { routePaths } from '@/constants/route';
 type MyActivityTemplateProps = {
   user: UserInfoType;
 };
