+++ src/components/templates/SignUpProfileLayout.template.tsx
@@ -2,7 +2,7 @@
 import { Children, ReactNode, useEffect, useRef, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useFormContext } from 'react-hook-form';
-import routePaths from '@/constants/routePaths';
+import { routePaths } from '@/constants/route';
 import Container from '@/components/atoms/Container';
 import IconButton from '@/components/molecules/IconButton';
 import Typography from '@/components/atoms/Typography';
