+++ src/components/templates/SignUpdatePassword.template.tsx
@@ -3,14 +3,13 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import { Link } from 'react-router-dom';
 import { useState } from 'react';
 
-
 import Container from '@/components/atoms/Container';
 import FormItem from '@/components/molecules/FormItem';
 import Button from '@/components/atoms/Button';
 import Typography from '@/components/atoms/Typography';
 import { SignUpdatePassword, SignUpdatePasswordType } from '@/types/auth.type';
 import { useSignUpdatePassword } from '@/hooks/useSignPasswordReset';
-import routePaths from '@/constants/routePaths';
+import { routePaths } from '@/constants/route';
 
 export default function SignUpdatePasswordTemplate() {
   const [showPassword, setShowPassword] = useState(false);
