+++ src/hooks/useMyAccount.ts
@@ -1,7 +1,7 @@
 import { useMutation } from '@tanstack/react-query';
 import { useNavigate } from 'react-router-dom';
 import { useSetRecoilState } from 'recoil';
-import routePaths from '@/constants/routePaths';
+import { routePaths } from '@/constants/route';
 import { supabase } from '@/libs/supabaseClient';
 import { AccountFormType } from '@/types/account.type';
 import { createToast, errorToast, successToast } from '@/libs/toast';
