+++ src/hooks/useSignPasswordReset.ts
@@ -1,6 +1,6 @@
 import { useMutation } from '@tanstack/react-query';
 import { useNavigate } from 'react-router-dom';
-import routePaths from '@/constants/routePaths';
+import { routePaths } from '@/constants/route';
 import {
   SignPasswordResetType,
   SignUpdatePasswordType,
